testcases.py with io.StringIO

+++ 1A.cpp
@@ -4,7 +4,6 @@
 int main(int argc, char const *argv[]) {
     double n, m, a;
     std::cin >> n >> m >> a;
-    std::cout << int(ceil(n / a) * ceil(m / a)) << std::endl;
-
+    std::cout << int(ceil(n / a)) * int(ceil(m / a)) << std::endl;
     return 0;
 }
